Added hard reset and some small fixes

+++ RANGING_LIB/src/ranging_wrapper.h
@@ -37,7 +37,7 @@ public:
     struct Ranging_Slave
     {
         Position position; // in degrees
-        unsigned long int address;
+        uint32_t address;
     };
     struct Lora_Device
     {
